requirements.txt
Flask
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate
